auth/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.6.5
asgiref==3.4.1; python_version >= '3.6'
asyncpg==0.23.0
click==8.0.1; python_version >= '3.6'
dnspython==2.1.0; python_version >= '3.6'
email-validator==1.1.3
fastapi==0.66.0
greenlet==1.1.0; python_version >= '3'
h11==0.12.0; python_version >= '3.6'
idna==3.2; python_version >= '3.5'
mako==1.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==2.0.1; python_version >= '3.6'
pipfile==0.0.2
psycopg2-binary==2.9.1
pydantic[email]==1.8.2
pyjwt==2.1.0
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.18.0
python-editor==1.0.4
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.20
starlette==0.14.2; python_version >= '3.6'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.10.0.0
uvicorn==0.14.0
